Implement commit watcher without GitHub API dependency
